wip: propose cid/get capability
At a high level, this capability is designed to enable us to display a list of spaces a CID's content has been uploaded to. There are a few complications:

1) a CID might identify a CAR (passed as the `link` of a `store/add` invocation or an actual "upload" (passed as the `root` of an `upload/add` invocation
2) a CID may have been uploaded to more than one space
3) a CID might even have been used in both `store/add` and `upload/add` - someone might have uploaded something and then re-uploaded the CAR that was passed to `store/add`

My proposal addresses all of these, returning a list of spaces a piece of content was `upload`ed to and a list of spaces a piece of content was `store`d in along with the dates that they were uploaded.

There are a bunch of other ways we might solve this, including:

1) separate capabilities for `upload/add`ed content and `store/add`ed content (i.e. `cid/get-upload` vs `cid/get-store`)
2) returning a single list of information with an additional `type` field to differentiate `store`d content from `upload`ed content
3) returning a single list with no differentiation
4) ??

If anyone has strong opinions on how this should work I'm definitely open to considering different designs!