feat: propose `root/get` and `shard/get` capabilities
At a high level, these are designed to enable administrators like @dchoi27 to figure out which space a given CID is associated with.

Problem CIDs come to us from a variety of sources - sometimes they are "root" CIDs that correspond to a particular upload and sometimes they are "shard" CIDs that identify one of the CARs that make up the shards of a larger upload. `root/get` returns information about the upload and `shard/get` returns information about particular shards, so given an unknown CID administrators should check both.