wip: propose `admin/blocks/inspect` capability
@dchoi27 often needs to get information about CIDs that identify blocks
packed into CARs stored. The current process for this is to
calculate the multihash of the CID (eg,
`ipfs cid format -f "%m" -b base58btc bafybeicfs2nuef675li6pxrdzar4bkmdp5ktud4shubtgkenk5gwtjqde4`)
and then to use that multihash to query the `blocks-cars-position` table,
which contains `carpath`, `length` and `offset`. He then uses the
`carpath` to determine the parent CAR's CID, which can be used to figure
out which space the content has been uploaded to, allowing him to block
the space and/or user who registered the space.

I'm very open to reworking this or tackling it a different way - this
is meant to be an initial proposal to start a discussion on the best
way to support this admin workflow.

Also tweak the return values of the `admin/store/inspect` and
`admin/upload/inspect` to reflect the (imho better) names used in the
implementations.