Revert provider revamp (#128)
# Goals

This PR reverts #107 

Why? The spec currently diverges wildly from the implementation, and
there are no immediate plans to implement the revised version.

We can absolutely return to this once we have immediate plans for
implementation. Generally, the specs should match exactly the
implementation, or there should be notes as to why.

I don't know the history and rationale for the revamp, but for now, it's
a huge source of confusion.