feat: `replicas` includes the original copy (#138)
Rework `space/blob/replicate` to count the original copy in the number
of replicas, and reword things to align with that.

Most notably, I've changed "Primary Node" to "Source Node" and
emphasized that it's not a property intrinsic to the node, but a
description of its role in the replication operation. Curious how that
looks to other folks.